화면 UI 만드는 중!!!

+++ yaKnifeObjectPool.cpp
@@ -31,7 +31,7 @@ namespace ya
 	}
 	UINT KnifeObjectPool::GetFreeObjectCount() const
 	{
-		return mQ.size();
+		return static_cast<UINT>(mQ.size());
 	}
 	UINT KnifeObjectPool::GetMaxFreeObjectCount() const
 	{
